requirements.txt
mysql-connector==2.2.9
numpy==1.17.4
pandas==0.25.3
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
